+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - added `filter` option as combined filter for `entryFilter` and `sourceFilter`
   - added `--all` option for CLI
   - removed default config file `.mcrrc.js` and `.mcrrc`
+  - added `find-up` for default config file 
   - fixed type style for UI
   - fixed string to multiple patterns "{...}" for minimatch filter
   - added README.zh-Hans.md